+++ src/irmin-pack/unix/inode.ml
@@ -68,4 +68,7 @@ struct
         in
         let v = Inter.Val.of_raw find v in
         Some v
+
+  let get_offset t k = Pack.get_offset t k
+  let get_length t k = Pack.get_length t k
 end